Refactor: Split monolithic main.py into modular architecture
- Split 1675-line main.py into 6 focused modules:
  - main.py: FastAPI app entry point and routing (139 lines)
  - models.py: Data models and classes (40 lines)
  - dependencies.py: Azure client management (112 lines)
  - logic_app_manager.py: Logic App concurrency management (217 lines)
  - blob_processing.py: Document processing pipeline (407 lines)
  - api_routes.py: All API endpoint handlers (635 lines)

- Benefits:
  - Improved maintainability and code organization
  - Better testability with isolated modules
  - Easier debugging and development
  - Clear separation of concerns

- Updated Dockerfile to use modular structure
- Added comprehensive test suite for validation
- Backed up original main.py as main_old.py
- Updated README with backend architecture documentation

- All 20 API endpoints preserved and functional
- No breaking changes to the API interface
